+++ examples/angular-app/src/app/pages/header.component.ts
@@ -1,0 +1,11 @@
+import { Component, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
+import { RouterLink } from '@angular/router';
+
+@Component({
+  selector: 'header-page',
+  templateUrl: './header.component.html',
+  standalone: true,
+  imports: [RouterLink],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA],
+})
+export class HeaderComponent {}