+++ examples/svelte-app/src/pages/Header.svelte
@@ -1,0 +1,7 @@
+<script>
+  import headerHtml from './html/header.html?raw';
+</script>
+
+<div>
+  {@html headerHtml}
+</div>