+++ packages/dev-kit-italia/elements.ts
@@ -19,6 +19,7 @@ import '@italia/chip/it-chip.js';
 // import '@italia/collapse/it-collapse.js';
 import '@italia/dropdown/it-dropdown-item.js';
 import '@italia/dropdown/it-dropdown.js';
+import '@italia/header/it-header.js';
 import '@italia/dimmer/it-dimmer.js';
 import '@italia/icon/it-icon.js';
 import '@italia/input/it-input.js';