+++ packages/dev-kit-italia/stories/components/alert.stories.ts
@@ -3,7 +3,7 @@ import { html } from 'lit';
 
 const meta = {
   title: 'Componenti/Alert',
-  tags: ['new', 'a11y-ok', 'documentation'],
+  tags: ['a11y-ok', 'documentation'],
   parameters: {
     layout: 'padded',
   },