+++ packages/breadcrumbs/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @italia/breadcrumbs
 
+## 1.0.0-alpha.9
+
+### Minor Changes
+
+- [#329](https://github.com/italia/dev-kit-italia/pull/329) [`0aa17c5`](https://github.com/italia/dev-kit-italia/commit/0aa17c545f8b0ba9d8e4b3550c0161c73f866148) Thanks [@deodorhunter](https://github.com/deodorhunter)! - Add bottom-nav component
+
+### Patch Changes
+
+- [`d12b0cb`](https://github.com/italia/dev-kit-italia/commit/d12b0cbcd0862320e601c24fd85d82ff7f7e657d) Thanks [@giuliaghisini](https://github.com/giuliaghisini)! - Fixed docs link to others components
+
+- Updated dependencies [[`b00d9b2`](https://github.com/italia/dev-kit-italia/commit/b00d9b262bec00d3a036e6a38ec8a0e4975d0474), [`fef3f8b`](https://github.com/italia/dev-kit-italia/commit/fef3f8bdde69cebe93c4b5d3b0fd75701f588188), [`a5ee60f`](https://github.com/italia/dev-kit-italia/commit/a5ee60f85c77c681a5c61a87970388318e794188)]:
+  - @italia/globals@1.0.0-alpha.9
+
 ## 1.0.0-alpha.8
 
 ### Patch Changes